+++ changelog.md
@@ -3,6 +3,7 @@
 ## 2.0.0
 
 - Now, AutoMessages are identified by either the Server IP, the Realm Name, or the World Name
+- Now, AutoMessage identifiers can be parsed as regex patterns
 - Now, AutoMessage delays are configured in decimal seconds
 - Fixed the caret position in regex syntax-error tooltips
 - Added a config option to enable command signing, fixing AutoMessage commands on certain servers